+++ ChangeLog.md
@@ -9,6 +9,8 @@ _2 September 2020_
 - `mv` command was failing ... added double quotes
 - For 4K forced-subtitling, changed `--add-subtitle "Forced"` to `--add-subtitle auto` so that the default subtitle and forced flags were auto-set.
 - Removed quotes from `--add-audio` and `--add-subtitle` strings to ensure both `call` commands and those run on the CLI act in the same way. Quotes lead to `--add-audio "AD"=surround` being processed as surround within a `call` command but as stereo (incorrectly) if run from the commandline.
+- Added in logic to handle `pcm_s24le` audio tracks
+
 
 ## 0.31 ##
 _28 August 2020_